Fix suspected 'canvasContect' typo
I am getting "ERROR TypeError: Cannot read property 'canvas' of undefined" when destroying a waveplayer instance.  I've traced it back to line 402 of WaveView.js.  It appears this may be a typo.  Is this supposed to be `canvasContext` rather than `canvasContect`?  If so, here is a pull request to fix the issue.